requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
pymupdf
transformers
requests
numpy 
python-dotenv
python-multipart